Added comments to the code
- In "Step 1: Create a new solidity file":
a) The interest rate is being stored as a percentage, which is not a good practice in Solidity, as it can cause rounding errors. It's better to store the interest rate as a decimal number, such as 0.05 for a 5% interest rate.
b) The code is vulnerable to reentrancy attacks in the `withdraw` function, as it allows the caller to execute arbitrary code in other contracts, which could be malicious. It's recommended to use the `withdraw pattern` to prevent reentrancy attacks, by setting the balance of the user to 0 before transferring the funds to avoid any recursive calls to the function.

- Applied code indentation.
-

-There are a few errors in "Step 4: Interact with the Smart Contract" which I have corrected for you:
a) ethers is not defined. You need to require the ethers library using `const ethers = require("ethers");` before using it.
b) `microlending.loans` returns a promise that needs to be awaited before accessing the loan object.
c) `microlending.repayLoan` needs to be passed the `repayAmount` as an argument.